Dodanie adaptera do odcinkow (w edit i remove)

+++ app/src/main/java/com/example/gotpttk/view/adminGui/spotsFragments/EditSpotFragment.java
@@ -75,7 +75,7 @@ public class EditSpotFragment extends Fragment
                         @Override
                         public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                             Bundle passId = new Bundle();
-                            passId.putLong("id", (int)id);
+                            passId.putLong("id", id);
                             EditChosenSpotFragment fragment = new EditChosenSpotFragment();
                             fragment.setArguments(passId);
                             getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment).commit();
